backend/requirements.txt
Django==3.2.3
djoser==2.1.0
djangorestframework==3.12.4
django-filter==23.1
django-extensions==3.1.5
django-link-shortener==0.5
environs==14.1.1
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
isort==5.13.2
Pillow==9.0.0
psycopg2-binary==2.9.3
requests~=2.32.3
webcolors==1.11.1